requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
chatterbox-tts==0.1.4